# perf(managed): avoid full checkout to validate local environment