@ysndr 2025-01-31
Benchmark 1: flox activate -- true
  Time (mean ± σ):     300.7 ms ±   5.9 ms    [User: 93.1 ms, System: 35.3 ms]
  Range (min … max):   293.1 ms … 315.2 ms    10 runs
 
Benchmark 2: /Users/yannik/.nix-profile/bin/flox activate -- true
  Time (mean ± σ):     411.0 ms ±  16.7 ms    [User: 56.5 ms, System: 33.6 ms]
  Range (min … max):   387.7 ms … 445.8 ms    10 runs
 
Summary
  flox activate -- true ran
    1.37 ± 0.06 times faster than /Users/yannik/.nix-profile/bin/flox activate -- true